+++ docs/changelog.rst
@@ -11,6 +11,8 @@ v0.15.0
 * feature: Add :opt:`normal-exit-codes` setting
 * feature: Add :opt:`resolv-conf` and :opt:`hosts-file` to sandbox config
 * feature: Add :opt:`bridged-network` option to sandbox config
+* breaking: By default ``/etc/hosts`` and ``/etc/resolv.conf`` will be mounted
+  if they are proper mount points (can be opt out in container config)
 
 
 .. _changelog 0.14.3: